+++ payments/RELEASES.md
@@ -108,3 +108,13 @@ Based on:
 - [typescript v2.3.1] payments
 ### Releases
 - [NPM v2.3.1] https://www.npmjs.com/package/@wingspan/payments/v/2.3.1 - payments
+
+## 2024-08-26 01:31:33
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.379.0 (2.404.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.4.0] payments
+### Releases
+- [NPM v2.4.0] https://www.npmjs.com/package/@wingspan/payments/v/2.4.0 - payments